move website to sites folder
